update/WK-1 - Update tests
