+++ src/test/java/com/github/blog/integration/PostReactionControllerIntegrationTest.java
@@ -1,4 +1,4 @@
-package com.github.blog.controller;
+package com.github.blog.integration;
 
 import com.github.blog.config.ContainerConfig;
 import org.junit.jupiter.api.BeforeEach;
@@ -27,9 +27,9 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
  */
 @Transactional
 @SpringBootTest(classes = ContainerConfig.class)
-@Sql(executionPhase = Sql.ExecutionPhase.BEFORE_TEST_CLASS, scripts = {"/db/insert-test-data-into-user-table.sql", "/db/insert-test-data-into-post-table.sql", "/db/insert-test-data-into-comment-table.sql", "/db/insert-test-data-into-comment_reaction-table.sql"})
+@Sql(executionPhase = Sql.ExecutionPhase.BEFORE_TEST_CLASS, scripts = {"/db/insert-test-data-into-user-table.sql", "/db/insert-test-data-into-post-table.sql", "/db/insert-test-data-into-post_reaction-table.sql"})
 @Sql(executionPhase = Sql.ExecutionPhase.AFTER_TEST_CLASS, scripts = "/db/clean-test-data.sql")
-public class CommentReactionControllerTests {
+public class PostReactionControllerIntegrationTest {
 
     private MockMvc mockMvc;
 
@@ -41,14 +41,14 @@ public class CommentReactionControllerTests {
     @Test
     @Rollback
     @WithUserDetails
-    @DisplayName("comment reaction controller: create")
-    void create_returnsCreatedCommentReactionDto_whenDataIsValid() throws Exception {
-        mockMvc.perform(post("/comment-reactions")
+    @DisplayName("post reaction controller: create")
+    void create_returnsCreatedPostReactionDto_whenDataIsValid() throws Exception {
+        mockMvc.perform(post("/post-reactions")
                         .contentType(MediaType.APPLICATION_JSON)
                         .content("""
                                 {
-                                "commentId": 2,
-                                "reactionId": 1
+                                "postId": 2,
+                                "reactionId": 2
                                 }
                                 """))
                 .andExpect(status().isOk())
@@ -56,14 +56,14 @@ public class CommentReactionControllerTests {
     }
 
     @Test
-    @WithUserDetails
-    @DisplayName("comment reaction controller: create - validation exception")
+    @WithUserDetails("kvossing0")
+    @DisplayName("post reaction controller: create - validation exception")
     void create_throwsExceptionForbidden_whenReactionAlreadyExists() throws Exception {
-        mockMvc.perform(post("/comment-reactions")
+        mockMvc.perform(post("/post-reactions")
                         .contentType(MediaType.APPLICATION_JSON)
                         .content("""
                                 {
-                                "commentId": 1,
+                                "postId": 1,
                                 "reactionId": 1
                                 }
                                 """))
@@ -72,10 +72,10 @@ public class CommentReactionControllerTests {
 
     @Test
     @Rollback
-    @WithUserDetails
-    @DisplayName("comment reaction controller: update")
-    void update_returnsUpdatedCommentReactionDto_whenDataIsValid() throws Exception {
-        mockMvc.perform(put("/comment-reactions/{id}", 1)
+    @WithUserDetails("kvossing0")
+    @DisplayName("post reaction controller: update")
+    void update_returnsUpdatedPostReactionDto_whenDataIsValid() throws Exception {
+        mockMvc.perform(put("/post-reactions/{id}", 1)
                         .contentType(MediaType.APPLICATION_JSON)
                         .content("""
                                 {
@@ -88,28 +88,28 @@ public class CommentReactionControllerTests {
 
     @Test
     @Rollback
-    @WithUserDetails
-    @DisplayName("comment reaction controller: delete")
-    void delete_returnsDeletedCommentReactionDto_whenDataIsValid() throws Exception {
-        mockMvc.perform(delete("/comment-reactions/{id}", 1))
+    @WithUserDetails("kvossing0")
+    @DisplayName("post reaction controller: delete")
+    void delete_returnsDeletedPostReactionDto_whenDataIsValid() throws Exception {
+        mockMvc.perform(delete("/post-reactions/{id}", 1))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.id").value(1));
     }
 
     @Test
     @WithUserDetails
-    @DisplayName("comment reaction controller: find by id")
-    void find_findsCommentReactionById_whenDataIsValid() throws Exception {
-        mockMvc.perform(get("/comment-reactions/{id}", 1))
+    @DisplayName("post reaction controller: find by id")
+    void find_findsPostReactionById_whenDataIsValid() throws Exception {
+        mockMvc.perform(get("/post-reactions/{id}", 1))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.id").value(1));
     }
 
     @Test
     @WithUserDetails
-    @DisplayName("comment reaction controller: find all")
-    void find_findsAllCommentReactions_whenDataIsValid() throws Exception {
-        mockMvc.perform(get("/comment-reactions"))
+    @DisplayName("post reaction controller: find all")
+    void find_findsAllPostReactions_whenDataIsValid() throws Exception {
+        mockMvc.perform(get("/post-reactions"))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.content.length()").value(2));
     }
