+++ src/test/java/com/github/blog/integration/TagControllerIntegrationTest.java
@@ -1,4 +1,4 @@
-package com.github.blog.controller;
+package com.github.blog.integration;
 
 import com.github.blog.config.ContainerConfig;
 import org.junit.jupiter.api.BeforeEach;
@@ -29,7 +29,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @SpringBootTest(classes = ContainerConfig.class)
 @Sql(executionPhase = Sql.ExecutionPhase.BEFORE_TEST_CLASS, scripts = {"/db/insert-test-data-into-user-table.sql"})
 @Sql(executionPhase = Sql.ExecutionPhase.AFTER_TEST_CLASS, scripts = "/db/clean-test-data.sql")
-public class ReactionControllerTests {
+public class TagControllerIntegrationTest {
 
     private MockMvc mockMvc;
 
@@ -41,9 +41,9 @@ public class ReactionControllerTests {
     @Test
     @Rollback
     @WithUserDetails("admin")
-    @DisplayName("reaction controller: create")
-    void create_returnsCreatedReactionDto_whenDataIsValid() throws Exception {
-        mockMvc.perform(post("/reactions")
+    @DisplayName("tag controller: create")
+    void create_returnsCreatedTagDto_whenDataIsValid() throws Exception {
+        mockMvc.perform(post("/tags")
                         .contentType(MediaType.APPLICATION_JSON)
                         .content("""
                                 {
@@ -56,9 +56,9 @@ public class ReactionControllerTests {
 
     @Test
     @WithUserDetails
-    @DisplayName("reaction controller: create - bad request exception")
+    @DisplayName("tag controller: create - bad request exception")
     void create_throwsExceptionForbidden_whenUserDoesntHaveRightRole() throws Exception {
-        mockMvc.perform(post("/reactions")
+        mockMvc.perform(post("/tags")
                         .contentType(MediaType.APPLICATION_JSON)
                         .content("""
                                 {
@@ -71,9 +71,9 @@ public class ReactionControllerTests {
     @Test
     @Rollback
     @WithUserDetails("admin")
-    @DisplayName("reaction controller: update")
-    void update_returnsUpdatedReactionDto_whenDataIsValid() throws Exception {
-        mockMvc.perform(put("/reactions/{id}", 1)
+    @DisplayName("tag controller: update")
+    void update_returnsUpdatedTagDto_whenDataIsValid() throws Exception {
+        mockMvc.perform(put("/tags/{id}", 1)
                         .contentType(MediaType.APPLICATION_JSON)
                         .content("""
                                 {
@@ -87,27 +87,28 @@ public class ReactionControllerTests {
     @Test
     @Rollback
     @WithUserDetails("admin")
-    @DisplayName("reaction controller: delete")
-    void delete_returnsDeletedReactionDto_whenDataIsValid() throws Exception {
-        mockMvc.perform(delete("/reactions/{id}", 1))
+    @DisplayName("tag controller: delete")
+    void delete_returnsDeletedTagDto_whenDataIsValid() throws Exception {
+        mockMvc.perform(delete("/tags/{id}", 1))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.id").value(1));
     }
 
     @Test
     @WithUserDetails("admin")
-    @DisplayName("reaction controller: find by id")
-    void find_findsReactionById_whenDataIsValid() throws Exception {
-        mockMvc.perform(get("/reactions/{id}", 2))
+    @DisplayName("tag controller: find by id")
+    void find_findsTagById_whenDataIsValid() throws Exception {
+        mockMvc.perform(get("/tags/{id}", 2))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.id").value(2));
     }
 
     @Test
     @WithUserDetails("admin")
-    @DisplayName("reaction controller: find all")
-    void find_findsAllReactionsByPageSizeAndPageNumber_whenDataIsValid() throws Exception {
-        mockMvc.perform(get("/reactions").param("pageSize", "1").param("pageNumber", "1"))
+    @DisplayName("tag controller: find all")
+    @Sql({"/db/insert-test-data-into-post-table.sql", "/db/insert-test-data-into-post_tag-table.sql"})
+    void find_findsAllTagsByPostId_whenDataIsValid() throws Exception {
+        mockMvc.perform(get("/tags").param("postId", "1"))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.content.length()").value(1));
     }
