--- src/test/java/com/github/blog/service/OrderServiceImplTests.java
@@ -1,271 +1,0 @@
-package com.github.blog.service;
-
-import com.github.blog.controller.dto.common.OrderDto;
-import com.github.blog.controller.dto.request.OrderRequest;
-import com.github.blog.controller.dto.request.PageableRequest;
-import com.github.blog.controller.dto.request.filter.OrderFilterRequest;
-import com.github.blog.controller.dto.response.PageResponse;
-import com.github.blog.model.Order;
-import com.github.blog.model.Post;
-import com.github.blog.model.User;
-import com.github.blog.repository.OrderRepository;
-import com.github.blog.repository.PostRepository;
-import com.github.blog.repository.UserRepository;
-import com.github.blog.repository.filter.OrderFilter;
-import com.github.blog.service.exception.ExceptionEnum;
-import com.github.blog.service.exception.impl.CustomException;
-import com.github.blog.service.impl.OrderServiceImpl;
-import com.github.blog.service.mapper.OrderMapper;
-import com.github.blog.service.statemachine.event.OrderEvent;
-import com.github.blog.service.statemachine.state.OrderState;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.domain.Specification;
-import org.springframework.statemachine.StateMachine;
-import org.springframework.statemachine.StateMachineEventResult;
-import org.springframework.statemachine.service.StateMachineService;
-import reactor.core.publisher.Flux;
-import reactor.core.publisher.Mono;
-
-import java.util.Collections;
-import java.util.Optional;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-
-/**
- * @author Raman Haurylau
- */
-@ExtendWith(MockitoExtension.class)
-public class OrderServiceImplTests {
-
-    @Mock
-    private UserRepository userRepository;
-
-    @Mock
-    private OrderRepository orderRepository;
-
-    @Mock
-    private PostRepository postRepository;
-
-    @Mock
-    private OrderMapper orderMapper;
-
-    @Mock
-    private StateMachineService<OrderState, OrderEvent> stateMachineService;
-
-    @InjectMocks
-    private OrderServiceImpl orderService;
-
-    private OrderRequest request;
-    private OrderDto returnedOrderDto;
-    private Order order;
-
-    private final Long id = 1L;
-    private final Long userId = 1L;
-    private final Long postId = 1L;
-
-    private final PageableRequest pageableRequest = new PageableRequest(10, null, "asc");
-    private final OrderFilterRequest orderFilterRequest = new OrderFilterRequest(null, null, null);
-
-    @BeforeEach
-    void setUp() {
-        User user = new User();
-        Post post = new Post();
-
-        request = new OrderRequest(postId, userId);
-
-        order = new Order();
-        order.setId(id);
-        order.setPost(post);
-        order.setUser(user);
-
-        returnedOrderDto = new OrderDto(id, postId, userId, null, null);
-    }
-
-    @Test
-    @DisplayName("order service: reserve")
-    @SuppressWarnings("unchecked")
-    void reserve_reservesOrder_whenDataIsValid() {
-        StateMachine<OrderState, OrderEvent> sm = mock(StateMachine.class);
-        StateMachineEventResult<OrderState, OrderEvent> smResult = mock(StateMachineEventResult.class);
-
-        when(orderRepository.findById(id)).thenReturn(Optional.of(order));
-        when(stateMachineService.acquireStateMachine(order.getId().toString())).thenReturn(sm);
-        when(sm.sendEvent(any(Mono.class))).thenReturn(Flux.just(smResult));
-        when(smResult.getResultType()).thenReturn(StateMachineEventResult.ResultType.ACCEPTED);
-        when(orderMapper.toDto(order)).thenReturn(returnedOrderDto);
-
-        OrderDto reservedOrderDto = orderService.reserve(id);
-
-        assertNotNull(reservedOrderDto);
-        assertEquals(id, reservedOrderDto.id());
-        assertEquals(postId, reservedOrderDto.postId());
-    }
-
-    @Test
-    @DisplayName("order service: reserve - state transition denied")
-    @SuppressWarnings("unchecked")
-    void reserve_throwsException_whenStateTransitionDenied() {
-        StateMachine<OrderState, OrderEvent> sm = mock(StateMachine.class);
-        StateMachineEventResult<OrderState, OrderEvent> smResult = mock(StateMachineEventResult.class);
-
-        when(orderRepository.findById(id)).thenReturn(Optional.of(order));
-        when(stateMachineService.acquireStateMachine(order.getId().toString())).thenReturn(sm);
-        when(sm.sendEvent(any(Mono.class))).thenReturn(Flux.just(smResult));
-        when(smResult.getResultType()).thenReturn(StateMachineEventResult.ResultType.DENIED);
-
-        CustomException exception = assertThrows(CustomException.class, () -> orderService.reserve(id));
-
-        assertEquals(ExceptionEnum.STATE_TRANSITION_EXCEPTION, exception.getExceptionEnum());
-    }
-
-    @Test
-    @DisplayName("order service: cancel")
-    @SuppressWarnings("unchecked")
-    void cancel_cancelsOrder_whenDataIsValid() {
-        StateMachine<OrderState, OrderEvent> sm = mock(StateMachine.class);
-        StateMachineEventResult<OrderState, OrderEvent> smResult = mock(StateMachineEventResult.class);
-
-        when(orderRepository.findById(id)).thenReturn(Optional.of(order));
-        when(stateMachineService.acquireStateMachine(order.getId().toString())).thenReturn(sm);
-        when(sm.sendEvent(any(Mono.class))).thenReturn(Flux.just(smResult));
-        when(smResult.getResultType()).thenReturn(StateMachineEventResult.ResultType.ACCEPTED);
-        when(orderMapper.toDto(order)).thenReturn(returnedOrderDto);
-
-        OrderDto cancelledOrderDto = orderService.cancel(id);
-
-        assertNotNull(cancelledOrderDto);
-        assertEquals(id, cancelledOrderDto.id());
-        assertEquals(postId, cancelledOrderDto.postId());
-    }
-
-    @Test
-    @DisplayName("order service: buy")
-    @SuppressWarnings("unchecked")
-    void buy_completesOrder_whenDataIsValid() {
-        StateMachine<OrderState, OrderEvent> sm = mock(StateMachine.class);
-        StateMachineEventResult<OrderState, OrderEvent> smResult = mock(StateMachineEventResult.class);
-
-        when(orderRepository.findById(id)).thenReturn(Optional.of(order));
-        when(stateMachineService.acquireStateMachine(order.getId().toString())).thenReturn(sm);
-        when(sm.sendEvent(any(Mono.class))).thenReturn(Flux.just(smResult));
-        when(smResult.getResultType()).thenReturn(StateMachineEventResult.ResultType.ACCEPTED);
-        when(orderMapper.toDto(order)).thenReturn(returnedOrderDto);
-
-        OrderDto boughtOrderDto = orderService.buy(id);
-
-        assertNotNull(boughtOrderDto);
-        assertEquals(id, boughtOrderDto.id());
-        assertEquals(postId, boughtOrderDto.postId());
-    }
-
-    @Test
-    @DisplayName("order service: find by id")
-    void findById_returnsOrderDto_whenIdIsValid() {
-        when(orderRepository.findById(id)).thenReturn(Optional.of(order));
-        when(orderMapper.toDto(order)).thenReturn(returnedOrderDto);
-
-        OrderDto foundOrderDto = orderService.findById(id);
-
-        assertNotNull(foundOrderDto);
-        assertEquals(id, foundOrderDto.id());
-        assertEquals(postId, foundOrderDto.postId());
-    }
-
-    @Test
-    @DisplayName("order service: find by id - not found exception")
-    void findById_throwsException_whenIdIsInvalid() {
-        when(orderRepository.findById(id)).thenReturn(Optional.empty());
-
-        CustomException exception = assertThrows(CustomException.class, () -> orderService.findById(id));
-
-        assertEquals(ExceptionEnum.ORDER_NOT_FOUND, exception.getExceptionEnum());
-    }
-
-    @Test
-    @SuppressWarnings("unchecked")
-    @DisplayName("order service: find all")
-    void findAll_returnsPageResponse_whenDataIsValid() {
-        OrderFilter filter = new OrderFilter();
-
-        Pageable pageable = PageRequest.of(pageableRequest.pageNumber(), pageableRequest.pageSize(), pageableRequest.getSort());
-        Page<Order> page = new PageImpl<>(Collections.singletonList(order), pageable, 1L);
-        PageResponse<OrderDto> pageResponse = new PageResponse<>(Collections.singletonList(returnedOrderDto), 1, 1, 0, 1);
-
-        when(orderMapper.toEntity(any(OrderFilterRequest.class))).thenReturn(filter);
-        when(orderRepository.findAll(any(Specification.class), any(Pageable.class))).thenReturn(page);
-        when(orderMapper.toDto(page)).thenReturn(pageResponse);
-
-        PageResponse<OrderDto> foundOrders = orderService.findAll(orderFilterRequest, pageableRequest);
-
-        assertNotNull(foundOrders);
-        assertEquals(1, foundOrders.content().size());
-        assertEquals(id, foundOrders.content().get(0).id());
-        assertEquals(postId, foundOrders.content().get(0).postId());
-    }
-
-    @Test
-    @SuppressWarnings("unchecked")
-    @DisplayName("order service: find all - not found exception")
-    void findAll_throwsException_whenNoOrdersFound() {
-        OrderFilter filter = new OrderFilter();
-
-        Pageable pageable = PageRequest.of(pageableRequest.pageNumber(), pageableRequest.pageSize(), pageableRequest.getSort());
-        Page<Order> page = new PageImpl<>(Collections.emptyList(), pageable, 1L);
-
-        when(orderMapper.toEntity(any(OrderFilterRequest.class))).thenReturn(filter);
-        when(orderRepository.findAll(any(Specification.class), any(Pageable.class))).thenReturn(page);
-
-        CustomException exception = assertThrows(CustomException.class, () -> orderService.findAll(orderFilterRequest, pageableRequest));
-
-        assertEquals(ExceptionEnum.ORDERS_NOT_FOUND, exception.getExceptionEnum());
-    }
-
-    @Test
-    @DisplayName("order service: update")
-    void update_returnsUpdatedOrderDto_whenDataIsValid() {
-        User user = new User();
-        Post post = new Post();
-
-        when(orderRepository.findById(id)).thenReturn(Optional.of(order));
-        when(userRepository.findById(userId)).thenReturn(Optional.of(user));
-        when(postRepository.findById(postId)).thenReturn(Optional.of(post));
-        when(orderMapper.toDto(order)).thenReturn(returnedOrderDto);
-
-        OrderDto updatedOrderDto = orderService.update(id, request);
-
-        assertNotNull(updatedOrderDto);
-        assertEquals(id, updatedOrderDto.id());
-        assertEquals(postId, updatedOrderDto.postId());
-    }
-
-    @Test
-    @DisplayName("order service: delete")
-    void delete_returnsDeletedOrderDto_whenIdIsValid() {
-        when(orderRepository.findById(id)).thenReturn(Optional.of(order));
-        when(orderMapper.toDto(order)).thenReturn(returnedOrderDto);
-
-        OrderDto deletedOrderDto = orderService.delete(id);
-
-        assertNotNull(deletedOrderDto);
-        assertEquals(id, deletedOrderDto.id());
-        assertEquals(postId, deletedOrderDto.postId());
-        verify(orderRepository, times(1)).delete(order);
-    }
-}
