+++ src/test/resources/db/insert-test-data-into-user_info-table.sql
@@ -1,7 +1,4 @@
-DO '
-DECLARE
-    user1_id bigint;
-    user2_id bigint; BEGIN SELECT id INTO user1_id FROM blog.user WHERE username = ''kvossing0'';
-SELECT id INTO user2_id FROM blog.user WHERE username = ''gmaccook1'';
-
-INSERT INTO blog.user_info (id, firstname, surname, university, major, company, job) VALUES (user1_id, ''Karl'', ''Doe'', ''Harvard University'', ''Computer Science'', ''Google'', ''Software Engineer''), (user2_id, ''Alice'', ''Johnson'', ''MIT'', ''Artificial Intelligence'', ''Facebook'', ''AI Researcher''); END' LANGUAGE plpgsql;
+INSERT INTO blog.user_info (id, firstname, surname, university, major, company, job)
+VALUES
+    (5, 'Karl', 'Doe', 'Harvard University', 'Computer Science', 'Google', 'Software Engineer'),
+    (6, 'Alice', 'Johnson', 'MIT', 'Artificial Intelligence', 'Facebook', 'AI Researcher');
